+++ src/Pim/Bundle/EnrichBundle/DependencyInjection/PimEnrichExtension.php
@@ -42,6 +42,7 @@ class PimEnrichExtension extends Extension
         $loader->load('entities.yml');
         $loader->load('event_listeners.yml');
         $loader->load('factories.yml');
+        $loader->load('files.yml');
         $loader->load('filters.yml');
         $loader->load('form_subscribers.yml');
         $loader->load('form_types.yml');